requirements.txt
discord.py==2.0.1
redis~=3.5.3
aiohttp>=3.7.4
requests~=2.24.0
SQLAlchemy~=1.3.18
python-dotenv==1.0.0
PyYAML~=5.3.1